Fix migration for travis

+++ db/migrate/20190828181707_add_datasetkey_occurrence.rb
@@ -1,11 +1,15 @@
 class AddDatasetkeyOccurrence < ActiveRecord::Migration[6.0]
   def up
-    add_column :occurrences, :datasetKey, :string, after: :gbifID, limit: 50
-    add_index :occurrences, :datasetKey
+    unless column_exists? :occurrences, :datasetKey
+      add_column :occurrences, :datasetKey, :string, after: :gbifID, limit: 50
+      add_index :occurrences, :datasetKey
+    end
   end
   
   def down
-    remove_column :occurrences, :datasetKey, :string, after: :gbifID, limit: 50
-    remove_index :occurrences, :datasetKey
+    if column_exists? :occurrences, :datasetKey
+      remove_column :occurrences, :datasetKey, :string, after: :gbifID, limit: 50
+      remove_index :occurrences, :datasetKey
+    end
   end
 end
